### 2023-01-06 #5048 @editorialbot
[PRE REVIEW]: pysr3: A Python Package for Sparse Relaxed Regularized Regression
**Submitting author:** @aksholokhov (<a href="http://orcid.org/0000-0001-8173-6236">Aleksei Sholokhov</a>)
**Repository:** https://github.com/aksholokhov/pysr3
**Branch with paper.md** (empty if default branch):
**Version:** v0.3.3
