### 2023-03-17 aksholokhov/pysr3#11 @blakeaw
JOSS Review - blakeaw
Hello,

This issue is part of the JOSS review at https://github.com/openjournals/joss-reviews/issues/5155. Here are my comments and questions (the sections mostly correspond to items in my checklist):

------

### General checks + Functionality

#### General checks: Data sharing and Reproducibility + Functionality: Functionality and Performance

Could you share the dataset (and code too, if applicable) used to generate the timing results reported in the table of Figure 1? Is it included in the files under paper/synthetic_data_4_1?

### Software paper

#### State of the field

Are there any other commonly used packages for regularized regression (using different methods)?

#### Figure 1

  * Figure 1 isn’t referenced in the text. Although its placement follows the text naturally, it may provide additional clarity to go ahead and add a reference in the preceding sentence.
  * The time measurements don’t specify units. I assume they are in seconds, but it might be good to include this detail in the figure or its caption.
  * It might be helpful to define the abbreviations/acronyms such as ALASSO and SCAD in the Figure caption since they are not defined in the text of the paper.
  * Here, Adaptive LASSO is abbreviated as ALASSO, but in the docs it is A-LASSO. It might be good to make them consistent.

#### References

Some references include their DOI and render with the DOI link in the pdf version. Just for consistency, I’d suggest including the DOIs for other references where possible.

### Documentation

#### Community guidelines

I didn’t see any specific guidelines for reporting issues/problems or seeking support in the docs or README. For example, I think you could point folks to the issues page of the GitHub repo and suggest they open a new issue for reporting issues/problems and seeking support.

Thanks,
Blake
