### 2023-03-17 #5155 @blakeaw
Hi @plaplant and @aksholokhov, thank you for your patience. I've been traveling the last couple of weeks, so I was a bit delayed in finishing up my review. I've completed my initial review with my comments at https://github.com/aksholokhov/pysr3/issues/11
