### 2023-03-20 #5155 @plaplant
@blakeaw thanks for your review! @aksholokhov please begin addressing the comments in the review, and reply to this issue when you feel that you've addressed them.

@mhu48 I see that you've filled out your review checklist. Do you have issues that you feel need to be addressed? If so, please make an issue on the upstream repository: https://github.com/aksholokhov/pysr3. If not, then please reply to this issue stating you feel it is ready for acceptance.
