### 2023-03-20 aksholokhov/pysr3#12 @mhu48
@mhu48 review
Minor revisions

1. Figure 1 includes too much information and too many subplots, making it difficult for the reader to quickly grasp the point. In particular, the font size is too small and hard to read.

2. Please consider adding more discussion in Section 4.2 to emphasize why the real data is relevant and what it can teach potential readers about implementing the proposed package.

3. Regarding the uploaded codes, better documentation for the sections is needed to help the community understand the organization.

Apart from these minor issues, the submission looks great. Thanks.

Ming
