### 2023-03-20 #5155 @mhu48
@plaplant Thanks for the update! I see I didn't post under the correct place earlier. The comments are submitted as a new issue under https://github.com/aksholokhov/pysr3/issues/12
