### 2023-03-21 6a3400c
Fix skmixed->pysr3 in MANIFEST.in
