+++ paper.bib
@@ -25,7 +25,8 @@ year = {2020}
   volume={7},
   pages={1404--1423},
   year={2018},
-  publisher={IEEE}
+  publisher={IEEE},
+  doi = {10.1109/ACCESS.2018.2886528}
 }
 
 @article{sholokhov2022relaxation,
@@ -49,7 +50,8 @@ year = {2020}
   number={22},
   pages={5811--5823},
   year={2019},
-  publisher={IEEE}
+  publisher={IEEE},
+  doi={10.1109/tsp.2019.2946029}
 }
 
 @article{murray2020global,
@@ -158,4 +160,28 @@ year = {2020}
   booktitle = {ECML PKDD Workshop: Languages for Data Mining and Machine Learning},
   year      = {2013},
   pages = {108--122},
+}
+
+@article{schelldorfer2011estimation,
+  title={Estimation for high-dimensional linear mixed-effects models using l1-penalization},
+  author={Schelldorfer, J{\"u}rg and B{\"u}hlmann, Peter and DE GEER, SARA VAN},
+  journal={Scandinavian Journal of Statistics},
+  volume={38},
+  number={2},
+  pages={197--214},
+  year={2011},
+  publisher={Wiley Online Library},
+  doi={10.1111/j.1467-9469.2011.00740.x}
+}
+
+@article{groll2014variable,
+  title={Variable selection for generalized linear mixed models by L 1-penalized estimation},
+  author={Groll, Andreas and Tutz, Gerhard},
+  journal={Statistics and Computing},
+  volume={24},
+  number={2},
+  pages={137--154},
+  year={2014},
+  publisher={Springer},
+  doi={10.1007/s11222-012-9359-z}
 }
