+++ paper.md
@@ -46,7 +46,7 @@ To date, the library supports linear models (classic regression) and linear mixe
    
 Reliable automated feature selection requires easy-to-use libraries, so that practitioners can test and compare multiple regularization approaches using their data, choose the most effective method, and apply the analysis at scale. These libraries should be flexible and modular enough to accommodate future developments, such as newly proposed regularizers, without forcing the practitioner to implement new solvers for each approach. At the same time, the libraries must be efficient and robust enough to handle common challenges such as ill-conditioned problems that arise in datasets with correlated predictors. The PySR3 library is designed to easily include new loss functions, constraints, information criteria, and regularization strategies. All PySR3 models fully adhere the standards and interface requirements of `sklearn` [@sklearn_api], providing a familiar interface to users. 
 
-Currently, PySR3 models can be used to automatically select features in both linear regression models and linear mixed effects (LME) models, which extend linear regression to clustered data settings. LME models commonly arise in longitudinal data analysis and meta-analysis. Feature selection for linear mixed-effects models is harder than for linear regression models due to  non-linearity of LMEs and within-group correlations. To the best of our knowledge, there are no standalone Python packages for mixed-effect feature selection, while the alternatives implemented in R, such as lmmlasso [@schelldorfer2011estimation] and glmmlasso [@groll2014variable], are take more time to converge and yield worse selection accuracy, as discussed in [@sholokhov2022relaxation].
+Currently, PySR3 models can be used to automatically select features in both linear regression models and linear mixed effects (LME) models, which extend linear regression to clustered data settings. LME models commonly arise in longitudinal data analysis and meta-analysis. Feature selection for linear mixed-effects models is harder than for linear regression models due to  non-linearity of LMEs and within-group correlations. To the best of our knowledge, there are no standalone Python packages for mixed-effect feature selection, while the alternatives implemented in R, such as lmmlasso [@schelldorfer2011estimation] and glmmlasso [@groll2014variable], take more time to converge and yield worse selection accuracy, as discussed in [@sholokhov2022relaxation].
 
 # Core idea and structure of `pysr3`
 
@@ -58,7 +58,7 @@ Each regularizer included in PySR3 can also be used in its relaxed SR3 form [@zh
 PGD on the SR3-transformed problem takes fewer iterations to converge, and the features selected 
 are more accurate and have lower false positive rates across simulated examples for both linear regression [@zheng2018unified] and LME [@sholokhov2022relaxation] models, as illustrated on Figure \autoref{fig:lme_summary}.
 
-![SR3 relaxation makes a problem more well-conditioned which accelerates the optimization.\label{fig:lme_summary}](images/summary_improved.jpg)
+![SR3 relaxation makes a problem more well-conditioned which accelerates the optimization.\label{fig:lme_summary}](images/summary_improved.png)
 
 More information about the structure of the library can be found in [documentation](https://aksholokhov.github.io/pysr3/), 
 while the mathematical contributions are extensively discussed in [@zheng2018unified] for linear regression 
