### 2023-03-28 31d04ff
Delete Figure
