+++ paper.bib
@@ -33,13 +33,15 @@ year = {2020}
   title={A Relaxation Approach to Feature Selection for Linear Mixed Effects Models},
   author={Sholokhov, Aleksei and Burke, James V and Santomauro, Damian F and Zheng, Peng and Aravkin, Aleksandr},
   journal={arXiv preprint arXiv:2205.06925},
-  year={2022}
+  year={2022},
+  doi={10.48550/arXiv.2205.06925}
 }
 @article{aravkin2022relaxationb,
   title={Analysis of Relaxation Methods for Feature Selection in Mixed Effects Models},
   author={Aravkin, Aleksandr and Burke, James and Sholokhov, Aleksei and Zheng, Peng},
   journal={arXiv preprint arXiv:2209.10575},
-  year={2022}
+  year={2022},
+  doi={10.48550/arXiv.2209.10575}
 }
 
 @article{baraldi2019basis,
@@ -62,7 +64,8 @@ year = {2020}
   number={10258},
   pages={1223--1249},
   year={2020},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/S0140-6736(20)30752-2},
 }
 
 
@@ -74,7 +77,8 @@ year = {2020}
   number={2},
   pages={460--477},
   year={2014},
-  publisher={Taylor \& Francis}
+  publisher={Taylor \& Francis},
+  doi={10.1080/10618600.2013.773239}
 }
 
 
@@ -83,7 +87,8 @@ year = {2020}
   author={Li, Xiaoping and Wang, Yadi and Ruiz, Rub{\'e}n},
   journal={IEEE transactions on cybernetics},
   year={2020},
-  publisher={IEEE}
+  publisher={IEEE},
+  doi={10.1109/TCYB.2020.2982445}
 }
 
 @article{miao2016survey,
@@ -93,7 +98,8 @@ year = {2020}
   volume={91},
   pages={919--926},
   year={2016},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.procs.2016.07.111}
 }
 
 @article{Zheng2020,
@@ -134,7 +140,8 @@ year = {2020}
   title={A Proof of Principle: Multi-Modality Radiotherapy Optimization},
   author={Levin, Roman and Aravkin, Aleksandr Y and Kim, Minsun},
   journal={arXiv preprint arXiv:1911.05182},
-  year={2019}
+  year={2019},
+  doi={10.48550/arXiv.1911.05182}
 }
 
 @article{scikit-learn,
@@ -146,7 +153,8 @@ year = {2020}
  journal={Journal of Machine Learning Research},
  volume={12},
  pages={2825--2830},
- year={2011}
+ year={2011},
+ doi={10.5555/1953048.2078195}
 }
 
 @inproceedings{sklearn_api,
@@ -160,6 +168,7 @@ year = {2020}
   booktitle = {ECML PKDD Workshop: Languages for Data Mining and Machine Learning},
   year      = {2013},
   pages = {108--122},
+  doi = {10.48550/arXiv.1309.0238}
 }
 
 @article{schelldorfer2011estimation,
@@ -172,16 +181,4 @@ year = {2020}
   year={2011},
   publisher={Wiley Online Library},
   doi={10.1111/j.1467-9469.2011.00740.x}
-}
-
-@article{groll2014variable,
-  title={Variable selection for generalized linear mixed models by L 1-penalized estimation},
-  author={Groll, Andreas and Tutz, Gerhard},
-  journal={Statistics and Computing},
-  volume={24},
-  number={2},
-  pages={137--154},
-  year={2014},
-  publisher={Springer},
-  doi={10.1007/s11222-012-9359-z}
 }
