### 2023-03-29 aksholokhov/pysr3#11 @aksholokhov
Hello,

First, thank you for a detailed and very actionable review. I hope that my responses below address your questions and concerns.

> Could you share the dataset (and code too, if applicable) used to generate the timing results reported in the table of Figure 1? Is it included in the files under paper/synthetic_data_4_1?

Of course. All datasets, code, and a complete reproducibility guide are available [here](https://github.com/aksholokhov/msr3-paper). The synthetic datasets are also saved in the directory that you mentioned.

> Are there any other commonly used packages for regularized regression (using different methods)?

There are, thank you for pointing it out. In [the bigger paper](https://arxiv.org/pdf/2205.06925.pdf) we conducted a comparison of the packages (Table 3), but we omitted it here due to the size constraints. I mentioned the alternative packages explicitly and added a reference to that study in lines 44-46.

> Figure 1 isn’t referenced in the text. Although its placement follows the text naturally, it may provide additional clarity to go ahead and add a reference in the preceding sentence.

Thank you for pointing it out. I added the reference to the figure in the preceding sentence, line 61.

> The time measurements don’t specify units. I assume they are in seconds, but it might be good to include this detail in the figure or its caption.

Right, the time is measured in seconds; I added the measurements to the table.

> It might be helpful to define the abbreviations/acronyms such as ALASSO and SCAD in the Figure caption since they are not defined in the text of the paper.

Right. I defined the abbreviations in lines 53-55.

> Here, Adaptive LASSO is abbreviated as ALASSO, but in the docs it is A-LASSO. It might be good to make them consistent.

Thank you. I ensured that they are consistent now within the paper.

> Some references include their DOI and render with the DOI link in the pdf version. Just for consistency, I’d suggest including the DOIs for other references where possible.

Fixed! All DOIs should be in place now.

> I didn’t see any specific guidelines for reporting issues/problems or seeking support in the docs or README. For example, I think you could point folks to the issues page of the GitHub repo and suggest they open a new issue for reporting issues/problems and seeking support.

The community guidelines are [here](https://aksholokhov.github.io/pysr3/community_guidelines.html); I added a note regarding the issues page on GitHub that you suggested.

Thank you again,

Aleksei

EDIT: added the link to the updated version of the paper. It's [here](https://github.com/aksholokhov/pysr3/actions/runs/4548862159).
