### 2023-03-29 aksholokhov/pysr3#12 @aksholokhov
Hello Ming,

Thank you for your review and I am glad to hear that you liked the submission overall. Below are my answers and questions regarding the revision points that you made. The updated version of the paper is [here](https://github.com/aksholokhov/pysr3/actions/runs/4548862159).

> Figure 1 includes too much information and too many subplots, making it difficult for the reader to quickly grasp the point. In particular, the font size is too small and hard to read.

Thank you for pointing it out. I agree that Figure 1 was hard to read, especially when printed. I reworked the figure per your recommendations and increased all font sizes to match those of the main text. I also improved the organization: now the sections of the figure are separated and the number of subplots is halved. Hopefully Figure 1 conveys the message better now and please let me know if it can be improved further.

> Please consider adding more discussion in Section 4.2 to emphasize why the real data is relevant and what it can teach potential readers about implementing the proposed package.

Just to double-check, do you refer to the JOSS paper (the paper under current review) from [here](https://github.com/openjournals/joss-reviews/issues/5155) or to the big paper from [here](https://arxiv.org/abs/2205.06925)? I am not sure if JOSS paper discusses the real-data application that the big paper discusses. We cite some applications to real-world data in the last section of the JOSS paper (Ongoing Research and Dissemination), but I am not sure if that discussion could be significantly expanded due to the size constrains.

> Regarding the uploaded codes, better documentation for the sections is needed to help the community understand the organization.

Currently the documentation is located [here](https://aksholokhov.github.io/pysr3/index.html), and the page that outlines the organization in tables and by functionality is [here](https://aksholokhov.github.io/pysr3/models_overview.html). We also have an auto-generated API documentation [here](https://aksholokhov.github.io/pysr3/source/modules.html#) that describes the modules and their content. Could you please give me some examples of documentation improvements that you would like to see?

Thank you again for your review and I am looking forward to hearing from you.

Aleksei
