### 2023-04-03 aksholokhov/pysr3#11 @blakeaw
Thank you for addressing my comments and clarifying my questions. I have a couple of minor follow-up suggestions:

1. Since the updated version of Figure 1 in the software paper is still pretty similar to Figure 1 of your bigger paper it may be good to add a note in the software paper figure caption stating something like that it is adapted from Figure 1 of Sholokhov 2022.
2. The msr3-paper repo was very helpful for further testing the code and reproducibility. I'd recommend adding a link to it somewhere in the pysr3 README or docs (or maybe in `pysr3/paper/readme.md`), as well as clarifying that it can also be used to recreate results/figures in Figure 1 of the software paper.

Thanks,
Blake
