### 2023-04-06 deed022
Edit JOSS paper. Add reference to the reproducibility guide. Version up.
