+++ src/pysr3/__about__.py
@@ -32,7 +32,7 @@ __classifiers__ = [
                       "Operating System :: OS Independent",
                   ],
 
-__version__ = "0.3.4"
+__version__ = "0.3.5"
 
 __author__ = "Aleksei Sholokhov"
 __email__ = "aksh@uw.edu"
