### 2023-04-06 aksholokhov/pysr3#11 @aksholokhov
Thank you for your suggestions, I incorporated both of them:
1. I added a mention just before the figure that it was adapted from Figure 1 of the bigger paper.
2. I added the link to the reproducibility guide on top of `pysr3/paper/readme.md`.

Thank you again for all your input.

Aleksei
