### 2023-04-06 aksholokhov/pysr3#11 @blakeaw
Great! You've addressed all my comments, so I'll go ahead and close this issue and ping the editor in the review thread (https://github.com/openjournals/joss-reviews/issues/5155) to let them know.
