### 2023-04-15 aksholokhov/pysr3#12 @mhu48
Hi Aleksei,

Sorry for the late reply!

I had some time to go over your new draft this morning and overall I am very happy with it. I feel your replies have sufficiently answered my questions and I can see you have also made other improvements in this version.

I think the documentation and the outline of the organization help a lot in understanding the structure, and I appreciate your other clarifications. I would suggest that this be accepted. Congratulations on your research!

Yours sincerely,

Ming
