### 2023-04-18 44694ca
Re-generate notebook using v0.3.5; add JOSS badge
