### 2023-04-18 abc13a8
Deploying to gh-pages from @ aksholokhov/pysr3@44694caf26cf7d5b0958499f4b9b568f68ba4d02 🚀
