### 2023-04-19 da0c7b7
Remove two references that are not cited in the paper.md
