+++ paper.bib
@@ -102,22 +102,6 @@ year = {2020}
   doi={10.1016/j.procs.2016.07.111}
 }
 
-@article{Zheng2020,
-abstract = {We develop and analyze a new 'relax-and-split' (RS) approach for inverse problems modeled using nonsmooth nonconvex optimization formulations. RS uses a relaxation technique together with partial minimization, and brings classic techniques including direct factorization, matrix decompositions, and fast iterative methods to bear on nonsmooth nonconvex problems. We also extend the approach to robustify any such inverse problem through trimming, a mechanism that robustifies inverse problems to measurement outliers. We then show practical performance of RS and trimmed RS (TRS) on a diverse set of problems, including: (1) phase retrieval, (2) semi-supervised classification, (3) stochastic shortest path problems, and (4) nonconvex clustering. RS/TRS are easy to implement, competitive with existing methods, and show promising results on difficult inverse problems with nonsmooth and nonconvex features.},
-author = {Zheng, Peng and Aravkin, Aleksandr},
-doi = {10.1088/1361-6420/aba417},
-file = {:Users/aksh/Documents/Papers/2020/Zheng, Aravkin/Relax-and-split method for nonconvex inverse problems/Zheng, Aravkin - 2020 - Relax-and-split method for nonconvex inverse problems.pdf:pdf},
-issn = {13616420},
-journal = {Inverse Problems},
-keywords = {nonconvex optimization,nonsmooth models,splitting methods},
-mendeley-groups = {Relax-and-Split examples},
-number = {9},
-publisher = {IOP Publishing},
-title = {{Relax-and-split method for nonconvex inverse problems}},
-volume = {36},
-year = {2020}
-}
-
 @article{Mendible2020,
 abstract = {We develop an unsupervised machine learning algorithm for the automated discovery and identification of traveling waves in spatiotemporal systems governed by partial differential equations (PDEs). Our method uses sparse regression and subspace clustering to robustly identify translational invariances that can be leveraged to build improved reduced-order models (ROMs). Invariances, whether translational or rotational, are well known to compromise the ability of ROMs to produce accurate and/or low-rank representations of the spatiotemporal dynamics. However, by discovering translations in a principled way, data can be shifted into a coordinate systems where quality, low-dimensional ROMs can be constructed. This approach can be used on either numerical or experimental data with or without knowledge of the governing equations. We demonstrate our method on a variety of PDEs of increasing difficulty, taken from the field of fluid dynamics, showing the efficacy and robustness of the proposed approach.},
 archivePrefix = {arXiv},
@@ -142,19 +126,6 @@ year = {2020}
   journal={arXiv preprint arXiv:1911.05182},
   year={2019},
   doi={10.48550/arXiv.1911.05182}
-}
-
-@article{scikit-learn,
- title={Scikit-learn: Machine Learning in {P}ython},
- author={Pedregosa, F. and Varoquaux, G. and Gramfort, A. and Michel, V.
-         and Thirion, B. and Grisel, O. and Blondel, M. and Prettenhofer, P.
-         and Weiss, R. and Dubourg, V. and Vanderplas, J. and Passos, A. and
-         Cournapeau, D. and Brucher, M. and Perrot, M. and Duchesnay, E.},
- journal={Journal of Machine Learning Research},
- volume={12},
- pages={2825--2830},
- year={2011},
- doi={10.5555/1953048.2078195}
 }
 
 @inproceedings{sklearn_api,
