### 2023-04-21 #5155 @danielskatz
An `@editorialbot` command needs to be the first thing in a comment
