### 2023-04-24 #5155 @aksholokhov
@arfon @plaplant Thank you so much! @saravkin, @ zhengp0, and I are thrilled that our paper is now published at JOSS and we are so grateful for all the work that you put into JOSS and in this review in particular.
@blakeaw @mhu48 Thanks a lot for your thorough and insightful reviews, they really helped me to improve the paper.

I am currently preparing for my graduation but starting June I would be happy to be on the list of possible reviewers for JOSS. It would be an honor for me to be a part of this community.
